Fix a bunch of minor portability problems and maybe-bugs revealed by running gcc and HP's cc with warnings cranked way up.  Signed vs unsigned comparisons, routines declared static and then defined not-static, that kind of thing.  Tedious, but perhaps useful...

+++ src/backend/libpq/hba.c
@@ -487,7 +487,7 @@ interpret_ident_response(char *ident_response,
 				cursor++;		/* skip blanks */
 			i = 0;
 			while (*cursor != ':' && *cursor != '\r' && !isblank(*cursor)
-				   && i < sizeof(response_type) - 1)
+				   && i < (int) (sizeof(response_type) - 1))
 				response_type[i++] = *cursor++;
 			response_type[i] = '\0';
 			while (isblank(*cursor))
